feat(remove): extracted remove.js to handle delte & remove confirmations
